mybatis整合redis作为数据库缓存

+++ src/main/java/com/anson/service/interfaces/IUserService.java
@@ -7,5 +7,5 @@ public interface IUserService {
 	
 	public String signUp(Users user);
 	
-	public int updateUserById(Users user);
+	public long updateUserById(Users user);
 }
